refactor: 커피챗 상태별로 컴포넌트 분리
